fixed last migration. added if statement for checking fields

+++ database/migrations/2020_09_04_172111_change_country_fields_at_all_tables.php
@@ -13,17 +13,31 @@ class ChangeCountryFieldsAtAllTables extends Migration
      */
     public function up()
     {
+        if (Schema::hasColumn('users', 'country'))
+        {
+            Schema::table('users', static function (Blueprint $table) {
+                $table->dropColumn('country');
+            });
+        }
         Schema::table('users', static function (Blueprint $table) {
-            $table->dropColumn('country');
             $table->bigInteger('country_id')->nullable();
-
         });
+        if (Schema::hasColumn('companies', 'country'))
+        {
+            Schema::table('companies', static function (Blueprint $table) {
+                $table->dropColumn('country');
+            });
+        }
         Schema::table('companies', static function (Blueprint $table) {
-            $table->dropColumn('country');
             $table->bigInteger('country_id')->nullable();
         });
+        if (Schema::hasColumn('clients', 'country'))
+        {
+            Schema::table('clients', static function (Blueprint $table) {
+                $table->dropColumn('country');
+            });
+        }
         Schema::table('clients', static function (Blueprint $table) {
-            $table->dropColumn('country');
             $table->bigInteger('country_id')->nullable();
         });
     }
